requirements.txt
numpy==1.19.5
scikit-learn==0.24.2
scikit-image
pillow==8.1.2
tensorflow-gpu==1.15.0
keras==2.4.3
pynput==1.7.3
h5py

scipy~=1.4.0
imageio==2.9.0
opencv-python~=4.5.3.0
mss
pyautogui
pydirectinput
